Remove time brackets and colotomic brackets
After reviewing the current state with my supervisor I took the
decision to remove the structure of colotomic brackets. It seems
to be unnecessarily complicated and too far from the usual situation
for interpreters of Western contemporary music. Furthermore I think
it makes the process of composition faster when using an (almost)
ordinary score.